GRACE_inference/grace-inference/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grace-inference"
version = "0.1.0"
description = "High-level Python library for GRACE machine learning force field inference tasks with graph basis functions"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "GRACE Inference Contributors"}
]
keywords = ["machine learning", "force field", "molecular dynamics", "materials science", "GRACE", "graph neural network", "MOF"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "grace-gnn>=0.1.0",
    "dgl>=1.0.0",
    "ase>=3.22.0",
    "phonopy>=2.20.0",
    "numpy>=1.21.0,<2.0.0",
    "scipy>=1.7.0",
    "h5py",
    "matplotlib>=3.5.0",
    "pandas",
    "tqdm",
    "prettytable",
    "pyyaml",
    "pymatgen>=2023.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.0.0",
    "dgl-cu118>=1.0.0",  # DGL with CUDA 11.8
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
all = [
    "grace-inference[gpu,dev]",
]

[project.urls]
Homepage = "https://github.com/yourusername/grace-inference"
Documentation = "https://github.com/yourusername/grace-inference#readme"
Repository = "https://github.com/yourusername/grace-inference.git"
Issues = "https://github.com/yourusername/grace-inference/issues"

[project.scripts]
grace-infer = "grace_inference.cli:main"

[tool.setuptools]
packages = ["grace_inference"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
grace_inference = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
